bib: tweak container architecture detection
This commit tweaks the way the architecuture is detected
for the bib containers. Instead of looking at the running
container config for the architecture and then trying
`uname -m` as fallback look for the container image and
then inspect that for the architecture. This seems to
be the most reliable way and work with the bib test
containers and with the ubi9 containers.