distro/generic: fix the version check for OSCAP in checkOptionsRhel8()
During a refactoring of the options checkers a long long time ago [1], I
accidentally overwrote the RHEL 8 version comparison with the one from
RHEL 9.  Before that, OpenSCAP on RHEL 8 was only supported for 8.7 and
above.
More recently, when dropping support for EOL releases, the config-map
changed and we started building RHEL 8.6 test images with openscap
options.

Fix the version check and update the config map to drop unsupported
versions of the test configs.

Add a test case for the error message in TestCheckOptions().

[1] e0182a9f935d100a041a848ef38c311f6149edca
[2] ab823cda069343b4e769d6446d37ca47d99c511c