distro/generic: remove tests that indirectly test checkOptions()
These tests would call the Manifest() function on all image types in a
distro and verify that it returns the expected error message for a bad
customization (or combinations).  This is now very redundant because:
- We only have one implementation of the imageType.Manifest() method
  (generic.imageType.Manifest()).
- The propagation of errors and warnings from checkOptions() through
  Manifest() is tested in TestManifestError().
- Error conditions from the checkOptions() function (and the
  distro-specific functions it calls) are tested in TestCheckOptions().

Repeating these tests in each distribution is now unnecessary, since
they all end up testing the same code paths.