distro: `hidden` image types
Some of our image types are 'internal use only', for example for RHEL
we tend to hide image types that are meant to be used with RHUI from
image type listings as RHUI content is not generally available on
premises.

The hiding of image types currently happens directly in
`osbuild-composer` [1] but we don't apply the same in `image-builder`.
By introducing a `hidden` flag on the image types (defaults to `false`)
we can filter out image types from listings while still letting them be
requested directly by users, or including them in listings when `-a` is
passed (or whatever flag we would like in the future, if we want this).

There are also other avenues to consider in the future that I'd be happy
to discuss (e.g. with externalized definitions we could *not* include
the definitions for the above image types in the vendored definitions).

[1]: https://github.com/osbuild/osbuild-composer/blob/2b8c5c4a595dc8b5cea04a48534916742696b837/cmd/osbuild-composer/config.go#L110-L117

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>