defs: hide image types
This hides image types based on the list currently being hidden in
`osbuild-composer` [1]. These images are not meant to be listed, only
built directly.

Note that we're not explicitly excluding `gce-rhui` as it's an alias.
For all others we're enabling `hidden` on the base image types that the
rest inherits from.

[1]: https://github.com/osbuild/osbuild-composer/blob/cbd63bf054ed117f617cbbc6475e3773cb389bb8/cmd/osbuild-composer/config.go#L110-L117

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>