tools: conflict-less manifest checksums
The current approach is overwriting manifest checksum files which causes
git conflicts. This patch appends last X characters of the checksum to the
filename to avoid conflicts.

While this doubles amount of git changes for each manifest update, it
prevents merge conflicts and is easier to handle. It also does not
increase the overall size of the repository because file overwrite is a
new blob for git anyway.