tools: fix gen-manifests-diff to support config changes
When gen-manifests-diff was written the use-case was to detect
the diff in code changes during the YAMLification. This lead
to the oversight that when a ./test/configs/* value or config-map
itself changes this was not displayed. This is confusing behavior
that this commit fixes.

Thanks to Achilleas for suggesting this.