disk: volume ID generation update
For our hardcoded partition tables we use hardcoded volume IDs for FAT
partitions. These are in a `FFFF-FFFF` format which is appropriate input
for the `fstab` stage. The `mkfs` stage converts this format to
something appropriate for the `mkfs` command.

However, when the `/boot/efi` partition is customized (which was allowed
in the previous commits) then a random volume ID is generated. This
random volume ID is not in the same format.

This causes the wrong volume ID to be written into `/etc/fstab` and thus
makes the system not boot as it awaits the volume to become available by
the identifier.

Fix up `NewVolIDFromRand` to return FAT volume IDs in the correct
format.

Thanks so much to Achilleas Koutsou who did the debugging [1] this
commit is based on. [1]

[1]: https://github.com/osbuild/images/pull/1792#issuecomment-3311563897

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>