disk: sort partitions in a partition table by their start position
The relayout() function until now would change the order of partitions
to put the root partition, volume group, or btrfs volume at the end, but
only did so by changing the start sector of partitions.  The order in
the array, which affects the creation order and determines the partition
number on the table, was never affected.  While this is generally valid,
it was surprising that the layout order did not match the partition
numbering.

This commit changes the relayout() function to always sort partitions in
the table by their start position.