osbuild: add pipeline name to monitor status
In order to print reasonable duration statistics, we need to know which
pipeline is currently running. The osbuild monitor JSON status messages
now include the pipeline name in the context object. Without this, one
could not tell if the slowest stage was RPM in os or build pipeline.

It also adds a testdata that is a complete and unmodified osbuild of a
Fedora 42 raw image.