manifest,osbuild: Change lorax templates to a list with branding
More than one lorax template can be run, and the runtime-cleanup.tmpl
will need to know about the logo and release branding packages so
support for that has been added to the lorax-script stage. This also
changes to using a list of templates to run, and it gets the template package name
from the yaml instead of using a bool to select between hardcoded
options (that wasn't hooked up, so it always used the 99-generic one).

Also bumps the osbuild version to v160 for lorax-script branding
support.