tools: drop `gen-bootc-diff` as it is no longer needed
We used `./tools/gen-bootc-diff` to ensure the manifest
generation for disk based bootc images between the `images`
library and the custom code in `bootc-image-builder` is
identical. Now that bootc-image-builder is using the images
library directly (c.f.
https://github.com/osbuild/bootc-image-builder/pull/1014)
this test is a tautology and not useful anymore.

We test the bootc based manifests generation now in images
via `./tests/bootc-fake-containers.yaml` and the manifests
and diff show up with our regular
`./tools/gen-manifest{,s}-{diff,checksums}` tooling.

If we move the rpm based bootc installer ISO generation into
images we will need to revert this removal again. But its
just a git revert away and there is no need to clutter the
repo with this no-longer-needed script.