distrodefs/rhel-9: remove dracut-config-generic from azure_base_pkgset
The package is already added by the platform packages, so it's redundant
to have it in the base package set as well.  This change has no effect
on generated manifests (mock or full).

This will allow us to share the azure_base_pkgset with the CVM image
type, which should not include any dracut packages.