test/scripts: add support for booting wsl on x86
This requires a batch script to import the image into WSL and run the
check script. A few assumptions are made: the user on the windows host
should be "azureuser", and all the required test scripts are uploaded to
"C:\Users\azureuser".

Skips fedora-41 as wsl-setup isn't available there, and it fails to
import.