test/configs: add jq to all test configs that require it
Add jq to the package list of each configuration that might need it.  jq
is now required by test/scripts/check-host-config.sh script in order to
parse each test config and verify the customizations.

The requirements are determined by the following criteria:
- The config is used in an image type that is booted in CI (AWS, Azure,
  and WSL).
- The config doesn't already install jq or something that includes it
  (for example, the core group on RHEL 9+ installs jq).

For a lot of image types, the empty-rhel config was replaced by the
jq-only-rhel config.  The empty-rhel config is now used for the net
installer.

A new config was added for RHEL 7 because it doesn't have jq available
and can't be included with the new jq-only-rhel config group.

Verified by generating all depsolved manifests and calling:

    grep "jq.*\.rpm" *ami* *ec2* *azure* *vhd* -L