test: drop usage of config variants for distros
Remove any usage of a test configuration variant for specific distros
that was added to work around the issue of not being able to apply the
same config file to multiple filter sets.

This primarily consists of the following configs being replaced by
'empty':
- ec2-sap-empty
- empty-rhel
- empty-fedora
- empty-rhel7

The following were also identical and have been replaced by a new config
called 'jq-only':
- empty-rhel-vhd
- jq-only-rhel

The 'partitioning-lvm-r8' config was identical to the 'partitioning-lvm'
config.

In addition, the 'partitioning-plain-r8' config, while slightly
different than 'partitioning-plain', is now applied to the RHEL 8 filter
set.  The non-r8 variant includes a custom /boot/efi partition.

Manifest checksums have changed because the rng seed depends on the
config name for each manifest.