distro/validator: treat empty slices as unset
Configs (Blueprints) can have initialised empty slices.  This can happen
if a user specifies the field but doesn't add any elements.  It can also
happen if an API is generating a request and initialises empty slices
even when they're not used, or serialises a request without (the
equivalent of) `omitempty` on a slice.  We should treat empty slices as
unset to avoid causing issues in such cases.