image: rename AnacondaContainerInstaller->A..C..Legacy
This commit renames the AnacondaContainerInstaller to
AnacondaContainerInstallerLegacy. The reason is that
we want to phase out this RPM based bootc container
image. It is only used in bootc-image-builder to
construct a boot ISO from the detected distros RPM
packages. This (arguably) goes against the spirit
of bootc where the container itself should be able
to provide the functionality of the installer and
not a bunch of RPMs.