many: add infrastructure for bootc anaconda images
This commit contains the infrastructure in the distro,
image and manifest packages to support building bootc
based ISO images. This includes:
- suppport for setting an installer payload
- a new bootc installer image type in pkg/distro/bootc
- support for containers in image/anaconda_container_intaller
- exported helper to share between bib/images until the legacy
  ISO rpm installer moved into images too (which is hard)

In addition (this was originally a second commit but
because the manifest checksum test run per-commit
this test update needs to be all one commit):

This commit adds reference manifests for the new bootc
iso installer image type. It also tweaks `gen-manifests`
a bit as we now need to split the tests between using
disk and installer image types.