github: drop automerge dependabot action
Using github.actor to determine if a PR was created by dependabot and
autoapprove it can be exploited using a "Confused Deputy" attack.

The automerge action is not useful anyway because we always need a
second approval, which requires human review.

Note also that dependabot is currently only configured to update GitHub
actions but it was, in the paste, configured to update Go dependencies.

See https://github.com/osbuild/images/security/advisories/GHSA-v669-m9gr-3mgp