disk: allow btrfs /boot
We enforce a partition for /boot even when there's already a subvolume
with a mountpoint for /boot in the partition table. Let's change this
requirement to only add a boot partition if there's a btrfs root but
*no* boot.

Note that in the future we might need to adjust this a bit further if we
plan to allow for setting parent subvolumes and automatic mountpoints
for btrfs subvols.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>