rhel-9: drop static UUIDs
Drop the static UUIDs defined in the distro definitions. Note that the
changes to the manifest checksums are only related to the PRNG now being
called more often and thus having a different state from before.

This also updates a UUID used in a test since the state of the PRNG is
different as it has generated *more UUIDs.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>