bootc: UEFI for aarch64
Currently we hard code the boot mode to LEGACY. This means that both a
BIOSBOOT and an ESP partition are required/created when missing. On
aarch64 this is always wrong and the BIOSBOOT should *not* be there.

This is a quick bandaid. In the future we want to share more and move
this into the same idiom as is being used for generic partitions.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>