disk: make addPartitionsForBootMode() arch specific
When building a disk from a disk customization blueprint we
will ensure that the right partitions for the boot mode of
the image is added. This code currently checks the boot-mode
of the image but for the aarch64/riscv64 architectures the
added partitions are subtly wrong if the inputs (i.e.
boot-mode) are wrong.

As a first step this commit makes addPartitionsForBootMode()
look at the architecture and silently discard boot modes
that do not make sense for the architecture. This is not
ideal and we really should instead error if we encounter
an invalid architecture+bootmode combination. But this
will fix the issue outlined in
https://github.com/osbuild/images/pull/1923 where we
create incorrect (x86) bios partitions on aarch64 bootc
images (and ricv64 too).