fedora/minimal: disk start offset for x86
For Fedora's ARM Minimal image we apply an offset to the x86 disk image.
This was introduced with the initial partition tables but is likely an
oversight. Let's drop the offset for x86 as I don't think it serves a
purpose there.

I've dropped these from all currently active branches instead of opting
for only-rawhide. I didn't see an easy-ish way to conditionalize this
without duplicating (part of) the partition tables.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>