disk: add new `disk.Offset` type for StartOffset
Offset describes the offset as a "size", this allows us to reuse
the datasize.Size json/toml/yaml loaders to specify the offset
as a nice string. We still make it a distinct type for readability
(as `StartOffset datasize.Size` would be weird).

Its not the most elegant way to go about this but should be okay
for now. Nicer would be to have an abstract type that has the
json/toml/yaml readers and then we use that to create
`datasizes.Size` and `disk.Offset` but finding a name for this
type is hard so this is probably okay for now (especially with
the comment for Offset).