distro: add `distro.ImageOptions.Bootc.InstallerPayloadRef`
Moving the installer payload ref for bootc into the
image options is a much nicer place than the previous
way of having it as part `BootcDistro`.

The alternative would be to make it part of
`BootcAnacondaInstaller`. This would imply an explicit
cast in all frontends so it seems this is preferable
(and similar to how ostree commits are configured).