distro: update bootc image types filenames
With the addition of the OVA image type the previous slightly
simplistic generation of the filename does not longer work and
we need to add the filename explicitely to the bootc image types.
This is a bit sad that we need to duplicate the data between
`pkg/distro/bootc/bootc.go` and `pkg/image/bootc_disk.go` that
also hardcodes the filenames. Eventually we should get back to
this.