fedora: disable `gpt-auto` for ISOs
Long bit of debugging went ahead of this, for full details see [1].

We want to inhibit `systemd-gpt-auto-generator` from running on ISOs
when they are booted in UEFI mode as it won't be able to find the root
partition. This in turn causes the boot to fail.

When `systemd-gpt-auto-generator` is prevented from running the boot
succeeds so this is likely an ordering issue in when/how things get
mounted by `dracut`.

This only affects (probably accidentally) non-live media as live-media
follows a different path to mount its root and probably runs a bit
earlier.

The solution here isn't a long-term thing though it's also not bad to
leave it in as there's never a need to run `systemd-gpt-auto-generator`
on ISOs as far as I'm aware.

[1]: https://github.com/osbuild/images/issues/1947#issuecomment-3395867961

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>