many: lorax templates as objects
Make the lorax templates in the image configs in the yaml definitions be
lists of a new type `InstallerLoraxTemplate` so that we can add
additional information to each template.

The additional information in this case is if the lorax template should
be executed before or after dracut. Since lorax templates often delete
things out of the filesystem tree it influences what dracut will copy
into the initramfs.

In a recent bug [1] this reared its head because an executable that was
previously deleted by lorax now ends up in the initramfs.

Since `runtime-cleanup.tmpl` deletes dracut it has to run after dracut
so let's move it there.

[1]: https://github.com/osbuild/images/issues/1947

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>