disk: use start offset customization from blueprint
Blueprint pull request https://github.com/osbuild/blueprint/pull/32 introduced a StartOffset DiskCustomization.
Use the value of this customization to adjust the start of the first partition in the partition table.

The StartOffset value specified in the blueprint will be added to the default offset which reserves space for the partition table (header).