many: lock root password
Give a simple image config option to lock the root password. In Fedora
43 the `setup` package now doesn't create an initial entry for the root
user and this entry is created by `systemd-sysusers`.

If there is no root user; or the root user contains the special value
`!unprovisioned` then `systemd-firstboot` triggers on the first boot of
the system to ask the user.

Some Fedora image types explicitly lock the root password. Let's have
the same functionality so we can use this.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>