Make bootupd metadata generation configurable per distro
Add a configurable option for bootupd metadata generation that allows
distributions to enable it only where the bootupd package is available.

This introduces the bootupd_gen_metadata field in ImageConfig, which
controls whether bootupctl backend generate-update-metadata is run for
ostree commits. Distributions can set this conditionally based on
version to ensure bootupctl is only invoked when bootupd is present.

Signed-off-by: Paul Whalen <pwhalen@fedoraproject.org>