gosbuild/osbuild-exec: add support for buildlog
This commit adds a new `BuildLog` option to the `OSBuildOptions` that
can be used to generate a streamed buildlog (e.g. to a file or a
websocket). This supports the `--with-buildlog` option in
image-builder-cli.

With the buildlog option set, stderr is forwarded to stdout to make sure
statements appear in order. If `BuildLogMu` is set, this mutex will be
used for the synced writer which writes to both stdout and the
buildlog. Callers can use the same mutex to write statements to the same
buildlog.

The errorWriter option got removed, and callers can just use `Stdout`
and `Stderr` to set the stdout and stderr of the command.