osbuild/monitor: increase status scanner buffer size to 16MiB
The size of the status message from the org.osbuild.rpm stage in the os
pipeline depends on how many packages a user has requested on top of the
base set. As a result, the monitor will always be breakable, but let's
increase the buffer size substantially to make it very unlikely that
users will hit this.

Especially with the addition of metadata [1], the status entry size grew
substantially for rpm stages. The rpm stage has an entry in the metadata
for each package that was installed (in addition to a very large output
message).

[1]: osbuild/osbuild#2215