data: introduce `data/distrodefs/shared/shared.yaml`
This commit adds a new `data/distrodefs/shared/shared.yaml` that
can be used to share yaml anchors between the various distros.

We have not that much duplicated data but some is the same so
de-duplicating it seems nice. The downside is that without the
move to `goccy/go-yaml` we need to add hackish code to make it
work [1].

[1] See https://github.com/osbuild/images/compare/main...mvo5:images:go-yaml-go
for the current showstoppers for this move.