distro: load distros from multiple files
Allow for splitting up `distros.yaml`. Any file directly at the top
level of the definitions filesystem is treated as a `distros.yaml` and
their entries are appended together.

This allows for slightly more organization for `distros.yaml`.

Note that these files are read in sorted filesystem order but separately
from eachother; hence anchors and such are not available across files
but only within themselves.

In the future this could/might lead to drop-in distro files and/or
merging of multiple filesystems.

We immediately make use of this new functionality by splitting up the
current single `distros.yaml` into `rhel.yaml` and `fedora.yaml`.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>