distrodefs: add new (hidden) bootc-rpm-installer imgtype
This commits adds a new `bootc-rpm-installer` image type
that contains only the rpm packages lists and installer
config for now so that bootc-image-builder can use the
images library to get what rpm package it needs to install
to build a bootable ISO.

The advantage is that we can retire the code in bib
that the YAML loading for the pacakge lists and that we
have one central location when we need to update image
definitions. The other advantage is that all the syntax
for conditions on the exact version works.

Note that this image type is not visible by default
because the loader will skip image types that have
no filename set.

This is used by https://github.com/osbuild/bootc-image-builder/pull/1066