reporegistry: refacor DistroHasRepos a bit
We have some ReposBy*() helpers already on RepoRegistry. So
make things conistent by renaming the existing DistroHasRepos()
to reposByDistroArch() as it behaves the same way as the others
so the naming should also be consistent.

Then create a compat wrapper for the old DistroHasRepos(). In
a followup we should either remove this compat wrapper (the
only consumer is weldr.go) or at least make it return eiher a
bool or an error, its weird that we have a "hasFoo" that returns
the value too.