reporegistry: fix TestInvalidReposByArch to work
This commit fixes the `TestInvalidReposByArch` which was not
actually testing much because with the existing code it would
error with:
```
failed to parse distro ID string: error when parsing distro name...
```

There was also a copy/paste mistake in 93c6336e24 which resulted
in the code testing the same code-path twice (i.e. both times
the "include-tagged" was set to false).

The test now tests for the real error message.