test: add support for local qemu testing via bibs vm.py
Initial support for running tests locally using boot-image
via the `vm.py` runner from bib for cloud-init supporting
images.

The root ssh key is added via cloud-init as this is what
our users are expecting (thanks to Ondrej for suggesting
this) and this way we do not need to modify the image itself.
This works for most images, for the others we need to discuss
(see below). This means we need "cloud-utils" installed.

We could also use test.thing but adding the polyfills is
actually not trivial and we need to modify the image. We could
also modify the image via guestfs and just add the root ssh
key directly, see mutate_qemu_image_for_login in [0]
for an example but the dowside of this is that we
modify the image (and its fiddly, we need to ensure
selinux is set correctly for anything we write via guestfs).

Currently only "qcow" is supported, cross arch boot is supported
for aarch64 (via the vmtest.vm.QEMU but it would be trivial
to extend this to more arches).

Note that because the QEMU test runner already supports cdrom
we can trivial test our installer with it too.

[0] https://github.com/osbuild/images/compare/main...mvo5:non-test-thing?expand=1