osbuild/tar: allow numeric owner
In a recent osbuild PR [1] an option was added to allow using numeric
owners in tar archives. This is wanted because including user and group
names in metadata can lead to unwanted side-effects when unpacking on a
system that contains user/group names.

Allow the tar stage to set this option in `images` as well.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>