bootc: add/switch to a non-running container by default
This commit switches our bootc introspection to not
actually run the container but only mount it. We need
this so that we can create osbuild manifests for bootc
containers without having to worry about attacks. A
container can potentially contain any code so its
risky to run it in our service.

We have the same problem when building (rpm) images where
the maintainer scripts can also run arbitrary code and we
run those in a secure instance. This is okay for image
building but quite expensive for manifest generation so
lets try to avoid it.

The one wringle here is that we can no longer run
```console
$ bootc install print-configuration
```
from the container. So this commit re-implements the
configuration reading from bootc. This is obviously
bad and we need to find a better solution.

One options is
```console
$ bootc install print-configuration --rootdir /path/to/container
```
so that we can run a trusted bootc to introspect the contaner
config. The downside of this of course is that if our bootc
is too old to understand new features from the container
(this is the beauty of running it from the container).

Another option might be to embrace this and accept that we
run arbitrary code and just trust that the podman isolation
will protect us. We will still need to harden our implementation
and ensure we run our podman container as a non-root user
(not sure this happens in the service currently).