test: add (basic) qemu based {image,minimal}-installer test
To run the tests manually do:
```console
$ ./test/scripts/build-image centos-10 image-installer test/configs/unattended-iso.json
$ OSBUILD_TEST_QEMU_GUI=1 ./test/scripts/boot-image centos-10 x86_64 image-installer ./build/centos_10-x86_64-image_installer-unattended_iso
```
note that the OSBUILD_TEST_QEMU_GUI=1 is optional but right now
without it there is just a lot of blank screen :) This is because anaconda
will display by default on the "last" console= from the kernel commandline
which is usually "tty0". For our tests ideally we would need to add a
blueprint with kernel.append="console=ttyS0".

This also add the "{image,minimal}-installer" to the list of
CAN_BOOT_TEST images.

Note that testing the "network-installer" requires us to make it
install with the `installer.unatteded: true` customization, otherwise
we will forever hang in the installer screen. This should be a
followup.