many: allow OCI archive options in image config
OCI archive options were hard coded in the various pipelines, or part of
the image types themselves. This commit moves them into a customization
under the OSCustomizations.

This moves visibility of these properties into the YAML and potentially
in the future will allow setting (some of these) properties through
blueprints.

Note that this commit does not actually switch the manifest generators
over yet. That happens next.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>