fedora: `container` labels and path
The Fedora `container` image type we had was deviated a bit as it was
lacking a PATH, entrypoint, and various other bits and bobs.

I inspected an official Fedora image and added what was missing to ours.
This now lets us run the image directly after a `podman load`.

We might still want to set `Author` and `Version`.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>