test: add support to check for server-qcow2 (no cloud-init)
This commit adds support for non-cloud init image types like
the `server-qcow2`. To test those we need to mutate the
image so that we can inject our test user and credetials
(note that in the future with test.thing and systemd
credentials passing this will be no longer necessary).

We also need a customization that contains the "osbuild"
user.

It can be run via:
```
./test/scripts/build-image fedora-43 server-qcow2 ./test/configs/all-customizations
./test/scripts/boot-image ./build/fedora_42-x86_64-server_qcow2
```