test: add support to check for minimal-root
This is a skeleton commit for the minimal-root image type. Here
we do not have cloud-init and with the empty image config
we have no user. So we need to modify the image to login.

It can be run via:
```
./test/scripts/build-image fedora-42 minimal-raw ./test/configs/empty.json
./test/scripts/boot-image fedora-42 x86_64 minimal-raw ./build/fedora_42-x86_64-minimal_raw-empty/
```

But its incomplete as is as the command is not transfered yet,
this needs a working scp in vm.py, see
https://github.com/osbuild/images/compare/main...mvo5:non-test-thing?expand=1