test: make runner image-type dependant
With the new qemu based testing of local images we want to make
sure that we pick gilab runners that support nested KVM. So this
commit adds a new `get_ci_runner_for(image_type)` helper that
allows to select specific runner for specific image types in
the Schutzfile. With that we can pick `rhos-01` which
supports kvm.