distro: add (legacy) rpm based anaconda-iso bootc installer
This commit moves the code to generate the rpm anaconda-iso
installer type into the images library. This is possible
because we use the distro specific depsolver. In bib we
need to define a custom depsolver function that wraps the
`manifestgen.DefaultDepsolver()` so that we can captures the
repositories and extract the mTLS keys from there.

The image type is ugly as it violates our layering, i.e.
it will first inspect the bootc container to find the distro
id and then look into our "regular" imagetypes.yaml for the
found distro for the rpms to install. This is convenient as
we maintain our rpms all in the same place but its also
annyoing. Ideally we would get rid of this image type and
use the container based bootc-installer type [0] but we
will probably need this RPM one for a while still.

[0] https://github.com/osbuild/images/pull/1906