distro: fix bootc default filesystem handling when missing
We need a way to
a) set the bootc default filesystem
b) detect when its unset and provide a clear error

This commit fixes both issue, we had a way to setup the default
filesystem but that is now applied too late because we now share
the partition table loading with the generic image types which
have no concept of an empty default. So this moves the default-fs
setting into the `bootc.NewBootcDistro()` as a new `DistroOptions`
struct.

It also adds a proper error type for when no default-fs is set,
this way ibcli can detect this and pick a default on its own
(this is what Achilleas suggested in [0]).

There is also a bugfix in setupDefaultFS() so that we do not
switch to "btrfs" on the /boot partition. This is the wrong
layer to know about /boot and details like this, we will
need a followup that moves this into disk instead.

[0] https://github.com/osbuild/image-builder-cli/pull/324#issuecomment-3455685910