customizations: `IsZero` for kickstarts
For the implementation of kickstarts in the network installer we want to
be able to know when a `kickstart.Options` is empty. This is so we can
add the kickstart only if it actually contains values that we need to
act upon instead of always (but with an empty kickstart file).

Note that this also adjusts the setting of users and groups from the
blueprint to use actual empty slices instead of slices of length 0. The
former is considered empty, the latter is not.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>