data: use yaml anchorto define "iso" bootc-generic image type
Ideally we would just use `name_aliases` but in the bib/ibcli
code we use the low-level interface of the loader to get the
image which does not look at name_alises. As this is quite
a special case lets just be pragmatic and define "iso" via
a YAML alias/anchor.