manifest: add treeinfo stage
Generated ISO files are missing .treeinfo file and thus it cannot be
used easily with virt-install/virt-manager/cockpit to install VMs. This
commit adds a new TreeinfoStage to osbuild that generates the .treeinfo
file based on the parameters passed from the anaconda installer ISO
manifest.

All fields supported by the new osbuild stage are defined, but one is
currently skipped: the ShortName because we do not have it in
distrodefs.