reporegistry: check for error before repo count
When an error occurs in loadAllRepositoriesFromFS(), the number of
returned repositories is always 0.  In LoadAllRepositories(), we check
for the number of loaded repos to return an error when no repositories
are loaded.  This means that when an error occurs while loading a
repository configuration, that error never bubbles up to the caller of
LoadAllRepositories().  Instead, it will always return
NoReposLoadedError, which makes troubleshooting very confusing.

Check for errors before verifying that the number of returned repos is
non-zero and return the underlying error if there is one.