test: switch boot-image in qemu to use the "osbuild" user
Instead of using the "root" user by default this commit
switches to the admin/sudo "osbuild" user. This is more
consistent with what the other clouds are doing and it
will also allow us to login into the `unattended-iso.json`
based images in the CI because they include the `osbuild`
ssh keys. With that we can run the real `check-host-config.sh`
script on the installed system after anaconda ran.