Schutzfile: increase rndseed to 9
The way we create qemu based cloud-inits has changed in the
previous commit. To ensure this has no unexpected bad side
effects re-run the full image tests. This is of course wasteful
on the non-qemu images which would not need to be boot tested
again but AFAIK there is no other way(?).