manifesgen: make "solver" mandatory in DepsolveFunc
This commit tries to address the concern from Achilleas in
https://github.com/osbuild/images/pull/2010#discussion_r2533915884

When using a custom Depsolve function it is now mandatory
to pass a valid `depsolvednf.Solver` instance. So the
custom depsolve function will no longer create it on its own
as this is no longer feasiable. With the bootc distro that
requires a different depsolvednf.Solver it is no longer
something that a user can override.

The commit also renames the function from Depsolver to
Depsolve as its doing the (dep) solving with an existing
"solver". Getting the terminology more disentangled would
be nice but I couldn't come up with better terms, maybe
`DepsolveRun` ?