gen-manifest: exclude image type
In this PR image types are getting renamed. Because the name of the
image type is part of the seed being used for the manifest checksums it
makes it unclear if the images are indeed only renamed as the checksums
change.

Let's try excluding the image type name from the seed so a rename should
still have the same checksum. This of course changes *all* checksums in
this commit; but subsequent commits should be more clear.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>