bootc-generic: Add a pxe-tar-xz image type to the bootc distro
Add the YAML definition for the new pxe-tar-xz image type.

The bootc container used for this has some requirements, a minimal
sample looks like this:

  FROM quay.io/centos-bootc/centos-bootc:stream10
  RUN dnf -y install dracut-live squashfs-tools erofs-utils && dnf clean all
  RUN bootc container lint

It needs dracut-live and the compression tools in order to generate a
new initramfs that can handle booting a compressed ostree filesystem. A
sample grub.cfg is included with the correct ostree=... kernel cmdline
setup for booting the included rootfs.img

NOTE: the bootc bootc-generic-growpart.service and
bootc-fetch-apply-updates.timer units are masked since they do not make
sense when booting from a ram based filesystem.

Resolves: HMS-9767
Related: HMS-8933